More gamemodes, improved target spawning
(Mostly) implemented the 4 current gamemodes, and reparented them to their respective classes (ie WideSpreadMultiBeat is child of MultiBeat, which is child of GameModeActorBase) and simplified the game mode selection/customization process in the main menu. The TargetSpawner now checks for collision dynamically based on the current amount of targets up at any given time using MaxTargetLifespan and TargetSpawnCD. Previously, it was only checking the area of the 4 most recently spawned targets.