AudioAnalyzer Customization added
Added AASettings, which is the widget that is contained inside the Settings menu (on main menu and pause menu). The player can now adjust how many band limits the analyzer uses, and what frequency each band uses. They can also change the threshold that determines whether or not to spawn a target for each band limit. The band limits require a game restart, so a "Save and Restart" button was added to pause menu". The thresholds can be changed at any time. Renamed BeatAimGameModeBase to DefaultGameMode. Deleted SpiderShotSelector.