Post-Game Screen, other improvements/fixes
Added a screen that appears after game mode has finished (PostGameMenuWidget) with options to restart, select a different game mode, and more. It also includes ScoreAnalysisWidget, which seems to work well besides reporting a bunch of stack errors to console. Made the gun its own actor and it is now the responsible for shooting projectiles. Moved some of the blueprint code to c++ for DefaultGameMode.