MuzzleFlash change, BeatTrack improvements
Muzzle flashes now spawn attached to the muzzle, and the muzzle flash is now local space instead of world space. This makes it so that the niagara system desync if moving and firing. BeatTrack targets now automatically start going in the opposite direction if they hit the edge of the spawn box. The direction of the BeatTrack target now takes into account the speed and cooldown of the target, so that it is much less likely to choose a direction where the target would exit the spawn bounds. Also fixed the linking of min/max target scale and min/max target speed so that entering a value in the text box while linked changes both the min and max values.