Game Difficulty, Menu changes, Dynamic settings
Added Game 3 game mode difficulties for the core 4 game modes. Changed the game modes user interface so that all game modes and difficulties are shown on one page, which then lets user customize or play with the defaults. Added spread type as a setting, which changes the spawn area based on performance. Dynamic target sizing behaves the same way.