Recoil and Fire Rate mostly finished
Finished implementing csgo-style recoil. Added bullet decals that spawn when a bullet collides with an object. Added fire rate and recoil settings, which can be changed by shooting the wall corresponding to the desired setting. Settings are automatically saved. Minor tuning adjustments to BeatTrack and BeatGrid. Added a streak tracker to the player hud which shows the max number of targets destroyed in a row for the current session. Changed the gun to spawn as a child actor to the character.