FPSCounter & Sensitivity tweaks, Framerate limit
Changed the fps counter so that it also appears in menus and in game. Changed the sensitivity slider to snap to values of 0.1. The new sens boxes now start empty and the current sens box now correctly shows the stored player's sensitivity. Added framerate limit options for the menu and in game.