GameModes change, login persist improvements, etc.
Made it so that single beat and multi beat only have one dynamic option since Edge only multibeat is essentially just singlebeat. Added ability to detect if the player scores menu is successfully showing the user's profile, and if not it asks them to log on again. Fixed a bug where the game wasn't sending the game mode difficulty to the database. Changed fade animations to delegates instead of using delays.