Improve login stuff, web browser fading
All in game web browser windows now fade in. If the screen requires authentication, it only fades in after it receives an http response. If it doesn't respond with a response code of 200, it doesn't show the web browser. The game checks if it can successfully load the player's profile page on launch, and if it fails it asks them to sign in again once, but doesn't ask more than that. The post game menu no longer shows the score web browser if the user has not signed in to the game yet.