Moved SphereTarget blueprint to C++
The changing of target colors is now done in C++, and only 3 total color curves are now required, instead of the 9 or so that were originally. This also allows makes room for the PlayerDelay to be a dynamic value instead of the 9 or so fixed delays that the player had to choose from.