Accuracy Data Stuff
Created new struct, FGridPoint to hold scaled spawned target coordinates along with a count. Created two arrays of FGridPoints to hold the number of spawns for each point and the number of  targets hit for each point. Wrote both to csv file so accuracy for every point can be easily attained.