More BeatGrid Options
Added separate sliders for the number of vertical and horizontal targets for BeatGrid. Finally implemented spacing between targets for BeatGrid (in the menu atleast). The max and min values of both the horizontal and vertical sliders are updated based on target scale, vertical & horizontal spread, and the number of vertical and horizontal targets. When spacing becomes constrained, a delegate is broadcasted to GameModesWidget with all the values that would relieve constraints. Still need to decides what priority is for relieving constraints, or just prompt user to pick. Still need to implement number of targets in TargetSpawner and target spacing, but should be simple.