Convert StaticCubeVisualizer to InstancedSM
The visualizers on the side of the spawn area are now created using instanced static meshes. Created 3 separate static meshes in Blender: one for the base cube, one for the outline on the sides of the cube that can be scaled with the cube, and the top outline mesh, that doesn't get scaled but only translated.