Ok now TrackGun ability actually fixed
Instead of always granting the TrackGun ability to the player (which just seems to get corrupted/garbage collected if it doesn't have the task running?), I opted to only grant the ability if the game mode uses the tracking damage type. If the user swaps game mode in the PostGameModeMenu to a non-tracking damage type, the TrackGunAbilityGrantedHandles are used to remove the AbilitySet from the player.