Fix ClusterBeat getting stuck
Fixed issue where the ClusterBeat mode could get stuck with one non-activated target. This stems from a larger problem that needs to be fixed for modes that allow spawn without activation. Need to have a check starting at top of GetValidSpawnLocations that trickles down to all the functions it calls that instructs whether to allow a managed target to be considered a valid spawn location