Gameplay Ability System Updates
Switched from using GameplayAttributeValueChangeDelegate to FBSAttributeEvent. Changed how damage is applied to targets: Instead of SetByCallerMagnitude, the HitDamage and TrackingDamage attributes are snapshotted from the source (player) and added together to modify the target's TotalDamage attribute, which then is subtracted from health in UBSAttributeSetBase::PostGameplayEffectExecute. Added BasePlayerHitDamage & BasePlayerTrackingDamage to BSConfig.