Patch 0.5.6 & Update CommonScoreInfo
CommonScoreInfo actually works now and stores a 5x5 accuracy matrix that contains the total spawns and total hits for each element. It also stores the QTable if using reinforcement learning along with the number of training samples that the table has experienced thus far. A CommonScoreInfo struct is unique for each Custom Game Mode or Preset Game Mode + Difficulty.

Still need to add a way to train the RLAgent even  if the user didn't select it in the settings. Also want to add a way to initialize the RLAgent from the accuracy matrix.