Http Request changes, steam stuff, Wall Menu
Modified all http requests to always take a shared pointer struct containing a callback delegate to avoid temporary declarations from being destroyed, and to allow for nested requests from temporary declarations. Moved score saving (to database) from BSGameMode to BSGameInstance so that it can happen asynchronously as the level changes if need be. Modified the Feedback widget so that it no longer clears feedback input upon an unsuccessful http request. Modified the quit menu so that the user is prompted when saving scores to database when exiting the desktop. Fixed an issue where multiple targets being spawned at the same time in non-grid modes could spawn intersecting each other. (line 1719 SpawnAreaManagerComponent). The Wall Menu now pulses the on or off text when changing the night mode setting until the time of day transition completes.