Bug fixes, various code improvements
Fixed a bug where starting a different game mode in the Post Game Mode Menu would not correctly load the new game mode and instead used the old game mode.
Fixed a bug where the QTable was not saved in game modes with reinforcement learning enabled.
Removed hardcoded level names to instead use level names set in the Game Instance blueprint.
Added functionality for applying equipped gameplay tags when a piece of equipment is equipped on an actor with an ability system component.
Made variables that are the same across all SpawnAreas static, and removed FSpawnAreaParams.
Removed functions and structs associated with a no longer used DFS method in SpawnAreaManagerComponent.